+++ karoo_gp/tree.py
@@ -11,6 +11,7 @@ class Tree:
 
     def __init__(self, id, root, tree_type='g', score=None):
         """Initialize a Tree from an id and Branch"""
+        # TODO: start from 0
         self.id = id        # The tree's position within population
         self.root = root    # The top Branch (depth = 0)
         self.tree_type = tree_type
