+++ karoo_gp/util/__init__.py
@@ -1,1 +1,1 @@
-from .lazy_loader import LazyLoader
+from .lazy_loader.lazy_loader import LazyLoader
