+++ karoo_gp/util/lazy_loader/README.md
@@ -1,0 +1,1 @@
+This directory contains the LazyLoader required to load Karoo quickly. It includes a copy of the original license, and the code is unchanged.
